Move base SVG support from kettle-ui to kettle-core for be able to access from engine for report generators
