+++ core/build.properties
@@ -7,3 +7,4 @@ impl.productID=kettle-core
 impl.version=${project.revision}
 
 dependency.pentaho-metastore.revision=TRUNK-SNAPSHOT
+dependency.batik.revision=1.7
