+++ core/src/org/pentaho/di/core/svg/SvgSupport.java
@@ -20,7 +20,7 @@
  *
  ******************************************************************************/
 
-package org.pentaho.di.ui.svg;
+package org.pentaho.di.core.svg;
 
 import java.io.InputStream;
