+++ ui/build.properties
@@ -11,4 +11,3 @@ dependency.kettle.revision=SVG-SNAPSHOT
 dependency.pentaho-xul.revision=TRUNK-SNAPSHOT
 dependency.pentaho-vfs-browser.revision=TRUNK-SNAPSHOT
 dependency.pentaho-reporting-core.revision=TRUNK-SNAPSHOT
-dependency.batik.revision=1.7
